Move seqcolapi into the refget package
