+++ refget/seqcolapi/__main__.py
@@ -1,0 +1,10 @@
+import sys
+
+from .main import main
+
+if __name__ == "__main__":
+    try:
+        sys.exit(main())
+    except KeyboardInterrupt:
+        print("Program canceled by user")
+        sys.exit(1)
