+++ refget/seqcolapi/static/index.html
