+++ seqcolapi/const.py
@@ -1,9 +1,7 @@
-import os
+"""Compatibility shim for :mod:`refget.seqcolapi.const`."""
 
-# ALL_VERSIONS moved into the distributed `refget` package (refget/const.py) so
-# the shared store-backed app factory can build service-info without seqcolapi,
-# which is not shipped in the refget wheel. Re-exported here for compatibility.
-from refget.const import ALL_VERSIONS  # noqa: F401
-
-STATIC_DIRNAME = "static"
-STATIC_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), STATIC_DIRNAME)
+from refget.seqcolapi.const import (  # noqa: F401
+    ALL_VERSIONS,
+    STATIC_DIRNAME,
+    STATIC_PATH,
+)
