+++ tests/integration/conftest.py
@@ -80,7 +80,7 @@ def loaded_dbagent(test_dbagent, test_fasta_path):
 def client(loaded_dbagent):
     """Create TestClient with test database"""
     from refget.router import get_dbagent
-    from seqcolapi.main import app
+    from refget.seqcolapi.main import app
 
     def override_get_dbagent():
         return loaded_dbagent
@@ -134,7 +134,7 @@ def test_server(request):
     import uvicorn
 
     from refget.router import get_dbagent
-    from seqcolapi.main import app
+    from refget.seqcolapi.main import app
 
     def override_get_dbagent():
         return loaded_dbagent
@@ -192,8 +192,8 @@ def store_test_server(tmp_path_factory):
     from fastapi.middleware.cors import CORSMiddleware
 
     from refget.router import create_refget_router, setup_backend
+    from refget.seqcolapi.const import ALL_VERSIONS
     from refget.store import RefgetStore
-    from seqcolapi.const import ALL_VERSIONS
 
     # Create store and load test FASTAs
     store_dir = tmp_path_factory.mktemp("store")
